feat: patch the math module only while a FlopCountingContext is active
Importing counted_float previously replaced 11 math functions process-wide at load time. The replacements now live in a dedicated module with refcounted apply/remove driven by FlopCountingContext enter/exit, so a plain import leaves the math module untouched and nested contexts unwind correctly. Version bumped to 1.1.0 (behavior change: math.* on a CountedFloat outside a context returns a plain float and counts nothing).
