+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "counted-float"
-version = "1.0.6"
+version = "1.1.0"
 description = "Count floating-point operations in Python code & benchmark relative flop costs."
 readme = "README.md"
 requires-python = ">=3.11"
